smc-monitoring/requirements.txt
#-e git+https://github.com/gabstopper/smc-python#egg=smc-python
file:../#egg=smc-python
websocket-client
